Changed network code to use MemoryReader in SimData
git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@1021 7152ba15-5bd8-0310-b828-e896570a432b

+++ CSP/CSPSim/Source/ClientNode.cpp
@@ -99,8 +99,8 @@ int ClientNode::run()
  // ptrPayload->timeStamp = 1.0;
 
   simdata::MemoryWriter writer((simdata::uint8 *)ptrPayload);
-  writer << (int)id;
-  writer << (int)type;
+  writer << id;
+  writer << type;
   writer << timestamp;
   
 //  b_GlobalPosition->value().writeBinary((unsigned char *)&(ptrPayload->globalPosition),24);
